AB#82810 - Move test helpers class so that it doesn't get picked up as a potential unit test candidate file
