+++ packages/layout/src/components/__tests__/employer.spec.tsx
@@ -4,7 +4,7 @@ import { EmployerCard } from '../cards/employer/employer';
 import { axe } from 'jest-axe';
 import userEvent from '@testing-library/user-event';
 import { Employer } from '../cards/employer/context';
-import { assertThatButtonHasBeenRemovedFromTheTabFlow } from './testHelpers';
+import { assertThatButtonHasBeenRemovedFromTheTabFlow } from '../testHelpers/testHelpers';
 
 const noop = () => Promise.resolve();
